sorting unchanged. Controller pre-starting on api methods
