+++ src/main/java/com/revature/services/JSONReaderService.java
@@ -1,4 +1,4 @@
-package com.revature.controllers;
+package com.revature.services;
 
 import java.io.FileNotFoundException;
 import java.io.FileReader;
@@ -10,7 +10,7 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
-public class JSONReaderController {
+public class JSONReaderService {
 
 //	@SuppressWarnings("unchecked") 
 	public static ArrayList<Object> dataCleaner(String[] args) {
