+++ src/main/resources/data.sql
@@ -1,0 +1,24 @@
+insert into batches values
+(1, 'Morgantown');
+insert into users values 
+(1, 'gpichmann0@artisteer.com', 'Grady', '5 Carpenter Plaza', 'New York City', 'NY', '10275', false, false, false, 'Pichmann', '212-374-3466', 'gpichmann0', '30401 Esker Point', 'Des Moines', 'IA', '50347', 1),
+(2, 'smigheli1@indiatimes.com', 'Salvidor', '6371 Elka Court', 'Columbus', 'OH', '43220', false, true, true, 'Migheli', '614-513-2776', 'smigheli1', '35747 3rd Park', 'Philadelphia', 'PA', '19131', 1),
+(3, 'fbraunroth2@ebay.co.uk', 'Fabien', '830 Pierstorff Parkway', 'Omaha', 'NE', '68197', false, true, true, 'Braunroth', '402-694-8099', 'fbraunroth2', '747 Crownhardt Place', 'New York City', 'NY', '10270', 1),
+(4, 'aocullen3@google.com.au', 'Aldon', '86978 Sage Junction', 'Elmira', 'NY', '14905', true, false, false, 'O''Cullen', '607-732-5313', 'aocullen3', '2 Killdeer Avenue', 'Norwalk', 'CT', '06859', 1),
+(5, 'hgledhill4@hatena.ne.jp', 'Hilda', '22514 Dryden Crossing', 'Naples', 'FL', '34114', true, true, true, 'Gledhill', '239-528-0279', 'hgledhill4', '45476 Oriole Drive', 'Winston Salem', 'NC', '27110', 1),
+(6, 'sdehoogh5@cnn.com', 'Simonne', '215 Twin Pines Plaza', 'Seattle', 'WA', '98115', false, false, false, 'De Hoogh', '206-506-8925', 'sdehoogh5', '68 Almo Terrace', 'New York City', 'NY', '10029', 1),
+(7, 'dosullivan6@loc.gov', 'Darren', '836 Saint Paul Trail', 'Columbus', 'OH', '43210', true, true, false, 'O Sullivan', '614-197-3998', 'dosullivan6', '8445 Novick Hill', 'Gainesville', 'GA', '30506', 1),
+(8, 'ddorro7@issuu.com', 'Donnell', '8 Erie Court', 'Fresno', 'CA', '93786', false, true, false, 'Dorro', '559-476-1222', 'ddorro7', '7175 Ruskin Point', 'Santa Monica', 'CA', '90405', 1),
+(9, 'bmacvean8@shop-pro.jp', 'Brynne', '2 Ridge Oak Plaza', 'Southfield', 'MI', '48076', true, false, false, 'MacVean', '734-143-5188', 'bmacvean8', '39545 Canary Trail', 'Durham', 'NC', '27710', 1),
+(10, 'tcurd9@illinois.edu', 'Tobe', '7565 Morningstar Avenue', 'Brooklyn', 'NY', '11205', false, true, true, 'Curd', '646-919-8954', 'tcurd9', '127 Dunning Plaza', 'Saint Louis', 'MO', '63131', 1),
+(11, 'medmundsa@tiny.cc', 'Marilyn', '736 Mcguire Drive', 'Los Angeles', 'CA', '90087', true, true, true, 'Edmunds', '213-600-8430', 'medmundsa', '6 Bay Road', 'Columbia', 'MO', '65218', 1),
+(12, 'wmatissoffb@facebook.com', 'Wynne', '4 Goodland Point', 'Oklahoma City', 'OK', '73167', false, false, true, 'Matissoff', '405-888-7795', 'wmatissoffb', '1 Lakeland Way', 'Shawnee Mission', 'KS', '66215', 1),
+(13, 'jlawfulc@simplemachines.org', 'Jozef', '31 Tomscot Hill', 'Spokane', 'WA', '99220', false, true, true, 'Lawful', '509-805-2290', 'jlawfulc', '492 Quincy Road', 'Houston', 'TX', '77035', 1),
+(14, 'flancashired@slideshare.net', 'Faythe', '20 Hermina Plaza', 'Cleveland', 'OH', '44105', true, true, true, 'Lancashire', '216-591-6593', 'flancashired', '1008 Pleasure Terrace', 'San Diego', 'CA', '92176', 1),
+(15, 'llynnitte@creativecommons.org', 'Lorita', '0889 Randy Way', 'Charlotte', 'NC', '28235', true, false, true, 'Lynnitt', '704-705-3599', 'llynnitte', '05579 Clarendon Avenue', 'Los Angeles', 'CA', '90015', 1),
+(16, 'modooghainef@reference.com', 'Marcellina', '300 Jenna Place', 'Boise', 'ID', '83757', true, true, false, 'O''Dooghaine', '208-387-2634', 'modooghainef', '27040 Carioca Crossing', 'Detroit', 'MI', '48242', 1),
+(17, 'jsellwoodg@wisc.edu', 'Jaine', '434 Pierstorff Hill', 'Wilmington', 'DE', '19892', false, true, false, 'Sellwood', '302-836-7801', 'jsellwoodg', '39 Merry Hill', 'Gadsden', 'AL', '35905', 1),
+(18, 'wviani@homestead.com', 'Wain', '4 Sommers Way', 'Charlotte', 'NC', '28210', false, false, true, 'Vian', '704-338-2790', 'wviani', '99 Delaware Hill', 'Athens', 'GA', '30605', 1),
+(19, 'gtootellj@latimes.com', 'Geneva', '27592 La Follette Trail', 'Anaheim', 'CA', '92812', false, false, false, 'Tootell', '714-167-0152', 'gtootellj', '61696 Algoma Park', 'Terre Haute', 'IN', '47812', 1),
+(20, 'cgeertzk@senate.gov', 'Cissy', '5 Spaight Court', 'Spring', 'TX', '77386', false, false, true, 'Geertz', '281-310-3238', 'cgeertzk', '3 Dexter Alley', 'Frederick', 'MD', '21705', 1),
+(21, 'lozintsevl@freewebs.com', 'Lynn', '26467 Loftsgordon Park', 'Charlotte', 'NC', '28242', true, false, true, 'Ozintsev', '704-108-6483', 'lozintsevl', '38 Muir Plaza', 'Pittsburgh', 'PA', '15210', 1);
